Update unit tests and fix lint errors
[#166734646]
