+++ src/material-ui-theme.tsx
@@ -41,10 +41,10 @@ export default createMuiTheme({
         padding: 14
       },
       thumb: {
-        width: 18,
-        height: 18,
-        boxShadow: "0 1px 5px 0 rgba(0, 0, 0, 0.35)",
-        border: "1px solid #797979",
+        "width": 18,
+        "height": 18,
+        "boxShadow": "0 1px 5px 0 rgba(0, 0, 0, 0.35)",
+        "border": "1px solid #797979",
         "$switchBase:hover &": {
           boxShadow: "0 0 0 3px rgba(255, 255, 255, 0.5)",
         },
@@ -53,11 +53,11 @@ export default createMuiTheme({
         }
       },
       switchBase: {
-        backgroundColor: 'transparent !important' // disable default hover state
+        backgroundColor: "transparent !important" // disable default hover state
       },
       track: {
-        backgroundColor: "#797979",
-        opacity: 1,
+        "backgroundColor": "#797979",
+        "opacity": 1,
         "$switchBase$checked + &": {
           opacity: 1
         }
